Reworked nextSlide() hook in BaseSlide.
Now the hook in BaseSlide always checks if the slide has defined any
states, and automatically cycles between them.
It assumes that all the states that are defined in the Slide are
sequential (ie. state1 comes before state2 and so on).

It's still possible to reimplement the hook to handle other cases.

Signed-off-by: Luca Ottaviano <lottaviano@develer.com>